skip-song-part 1.0.0
First public release with working multiple segments per song support and a user menu to easily view and remove segments

Moved most utils function to utils.tsx

Changed the bind to alt+s

Fixed a bug that made the extension only monitor the song in the context it was added. For example if you added segments to a song in the album then the exact same song would not be monitored if you played it from a playlist
Changed track uid to uri instead.

Added notification when starting a segment for better user experience

Created the user menu to view and delete all the saved segments.
Created menu.tsx

Added a README

Modified .gitattributes for release